Update Cypress version in project dependencies
This update involves transitioning from Cypress version 13.6.2 to
version 13.6.3 in both package-lock.json and package.json files. It's
important to ensure our test suite is using the latest stable release of
Cypress for better stability, potential fixes, and possible new
features. Also, as a part of this update, the dependency "@types/node"
was removed from the Cypress dependencies in package-lock.json. Reason
and impact of this change is unknown, pending investigation.